Add a test for Hosts.Add
We needed to check if the file was ending with a \n. In this case it was
not and this was causing problems with dnsmasq.

Writing tests in go seems to be quite messy. Perhaps there are a few
frameworks to make this simpler, but for now .. ewww. Specially in
comparison to mocha and the information it provides.

Plus, why do I need to write a custom error message each time! These two
values are not equal, that is the error message!